Move Android audio processing options to audio API
